requirements.txt
torch
torchvision
opencv-python
numpy
scipy
pillow
streamlit
transformers
langchain
